add test for #79, refactor to make test possible
